requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Deprecated==1.3.1
fastapi==0.121.0
h11==0.16.0
idna==3.11
img2pdf==0.6.3
iniconfig==2.3.0
lxml==6.0.2
numpy==2.2.6
ocrmypdf==16.4.1
opencv-python==4.12.0.88
packaging==25.0
pikepdf==8.10.1
pillow==12.0.0
pluggy==1.6.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytesseract==0.3.13
pytest==8.4.2
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wrapt==2.0.1
httpx==0.28.1
